+++ packages/icon/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gluestack-ui/icon
 
+## 0.1.4
+
+### Patch Changes
+
+- feat: migration of hook to react native aria
+- Updated dependencies
+  - @gluestack-ui/provider@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes