+++ packages/tooltip/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gluestack-ui/tooltip
 
+## 0.1.8
+
+### Patch Changes
+
+- feat: migration of hook to react native aria
+- Updated dependencies
+  - @gluestack-ui/overlay@0.1.5
+
 ## 0.1.7
 
 ### Patch Changes